chore(user): move files
