+++ src/user/user.type.ts
